Time: 5 ms (32.16%), Space: 69.6 MB (63.1%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -5,11 +5,11 @@ class Solution {
         for(int i=0;i<k;i++){
             sum+=nums[i];
         }
-        avg=sum/(float)k;
+        avg=sum/k;
         for(int i=1;i<nums.length-k+1;i++){
             sum-=nums[i-1];
             sum+=nums[i+k-1];
-            avg=Math.max(avg,sum/(float)k);
+            avg=Math.max(avg,sum/k);
         }
         return avg;
     }
